days of weeks from byte to int, package got image url and version, fix mathconcept

+++ MathBridgeSystem.Application/Services/ContractPdfGenerator.cs
@@ -114,7 +114,7 @@ namespace MathBridgeSystem.Application.Services
             }).GeneratePdf();
         }
 
-        private static string FormatDaysOfWeek(byte? daysOfWeek)
+        private static string FormatDaysOfWeek(int? daysOfWeek)
         {
             if (!daysOfWeek.HasValue || daysOfWeek == 0) return "No schedule";
